+++ vue-front-end/src/components/admin/CelestialBodies.vue
@@ -88,7 +88,7 @@
             Your search for "{{ search }}" found no results.
         </v-alert>  
         </v-data-table> 
-        <create-celestial-body :success="success" :formErrors="form" v-on:resetForm="resetForm($event)" v-on:sendValues="recieveValues($event)"></create-celestial-body>
+        <create-celestial-body :isUpdate="isUpdate" :updateFormVals="updateFormVals" v-on:sendValues="updateBodies($event)" v-model="openForm"></create-celestial-body>
         </v-card>
         <v-container>
             <div class="text-xs-center">
@@ -112,9 +112,6 @@
             </div>
         </v-container> 
     </div>
- 
-
-
 </template>
 <script>
 import router from '../../router';
@@ -128,41 +125,18 @@ export default {
     name: 'CelestialBodies',
     data(){
         return{
-            form: {
-                name: {
-                   value: "",
-                   hasError: false,
-                   errorMsg: ""
-               },
-               hours:{
-                   value: '',
-                   hasError: false,
-                   errorMsg: ""
-               },
-               minutes:{
-                   value: '',
-                   hasError: false,
-                   errorMsg: ""
-               },
-               seconds:{
-                   value: '',
-                   hasError: false,
-                   errorMsg: ""
-               },
-               declination:{
-                   value: '',
-                   hasError: false,
-                   errorMsg: ""
-               }
-            },
-            errorsList: {
+            updateFormVals: {
                 name: "",
                 hour: "",
                 min: "",
                 sec: "",
-                dec: ""
+                dec: "",
+                id: ""
             },
+            openForm: false,
             success: true,
+            callCreate: false,
+            isUpdate: false,
             data: {
                
             },
@@ -196,71 +170,8 @@ export default {
         }
     },
     methods: {
-        create(){
-            let data = JSON.stringify({
-                name: this.form.name.value,
-                hours: this.form.hours.value,
-                minutes: this.form.minutes.value,
-                seconds: this.form.seconds.value,
-                declination: this.form.declination.value
-            })
-
-            ApiDriver.CelestialBodies.createCB(data).then((response) => {
-                let that = this;
-               HttpResponse.then(response, data => {
-                    //this.$store.commit("loading", false);
-                    if(data.statusCode === "200"){
-                        success = true
-                    }
-                },(status, errors) => {
-                    //Populates error message for form, clears form if no errors are fixed
-                    if(errors.DECLINATION){
-                        this.form.declination.hasError = true
-                        for(var index in errors.DECLINATION){
-                            this.form.declination.errorMsg = this.form.declination.errorMsg + errors.DECLINATION[index] + "\n"
-                        }
-                    } else{this.form.declination.errorMsg = ""}
-                    if(errors.HOURS){
-                        this.form.hours.hasError = true
-                        for(var index in errors.HOURS){
-                            this.form.hours.errorMsg = this.form.hours.errorMsg + errors.HOURS[index] + "\n"
-                        }
-                    } else{this.form.hours.errorMsg = ""}
-                    if(errors.MINUTES){
-                        this.form.minutes.hasError = true
-                        for(var index in errors.MINUTES){
-                            this.form.minutes.errorMsg = this.form.minutes.errorMsg + errors.MINUTES[index] + "\n"
-                        }
-                    } else{this.form.minutes.errorMsg = ""}
-                    if(errors.SECONDS){
-                        this.form.seconds.hasError = true
-                        for(var index in errors.SECONDS){
-                            this.form.seconds.errorMsg = this.form.seconds.errorMsg + errors.SECONDS[index] + "\n"
-                        }
-                    } else{this.form.seconds.errorMsg = ""}
-                    if(errors.NAME){
-                        this.form.name.hasError = true
-                        for(var index in errors.NAME){
-                            this.form.name.errorMsg = this.form.name.errorMsg + errors.NAME[index] + "\n"
-                        }
-                    } else{this.form.name.errorMsg = ""}
-                    console.log(errors)
-                    that.handleErrors(errors)
-                }) 
-            }).catch((error) => {
-                console.log(error)
-                this.$swal({
-                            title: '<span style="color:#f0ead6">Error!<span>',
-                            html: '<span style="color:#f0ead6">An error occurred when loading the celestial bodies list<span>',
-                            type: 'error',
-                            background: '#302f2f'
-                        }).then(response => {
-                            CurrentUserValidation.validateCurrentUser(this.$store);
-                        });
-            });
-            this.success = false
-        },
         getCelestialBodies(){
+            this.bodies = []
             this.$store.commit("loading", true);
              ApiDriver.CelestialBodies.getCBList(this.pageNumber,this.selectedPageSize).then((response) => {
                 HttpResponse.then(response, data => {
@@ -289,32 +200,6 @@ export default {
             this.numPages = data.totalPages;
             //console.log(this.array)   
         },
-        handleErrors(errors) {
-          // Populate the field error messages
-          for (var field in errors) {
-              let message = errors[field][0];
-
-              if (field === "NAME") {
-                  CustomErrorHandler.populateError(this.form.name, message);
-              } else if (field === "HOURS") {
-                  CustomErrorHandler.populateError(this.form.hours, message);
-              } else if (field === "MINUTES") {
-                  CustomErrorHandler.populateError(this.form.minutes, message);
-              } else if (field === "SECONDS") {
-                  CustomErrorHandler.populateError(this.form.seconds, message);
-              }else if (field === "DECLINATION") {
-                  CustomErrorHandler.populateError(this.form.declination, message);
-              }  
-          }
-      },
-      clearErrors() {
-          // Clear all error fields
-          CustomErrorHandler.clearError(this.form.name);
-          CustomErrorHandler.clearError(this.form.declination);
-          CustomErrorHandler.clearError(this.form.hours);
-          CustomErrorHandler.clearError(this.form.minutes);
-          CustomErrorHandler.clearError(this.form.seconds);
-      },
         next(page) {
             // Handle retrieving a new page of information
             this.pageNumber = page - 1;
@@ -326,38 +211,21 @@ export default {
             this.bodies = []
             this.getCelestialBodies()
         },
-        recieveValues: function(vals){
-            this.form.name.value = vals.name
-            this.form.declination.value = vals.dec
-            this.form.hours.value = vals.hour
-            this.form.minutes.value = vals.min
-            this.form.seconds.value = vals.sec
-
+        updateForm(name,id,dec,hours,min,sec){
+            console.log("this is the CB id" + id)
+            this.isUpdate=!this.isUpdate
+            this.updateFormVals.name = name
+            this.updateFormVals.dec = dec
+            this.updateFormVals.hour = hours
+            this.updateFormVals.min = min
+            this.updateFormVals.sec = sec
+            this.updateFormVals.id = id
             
-            this.create()
         },
-        updateForm(name,id,dec,hours,min,sec,){
-            console.log(name)
-        },
-        resetForm: function(){
-
-            this.clearErrors()
-
-            this.form.name.value = ""
-            this.form.name.hasError = false
-            console.log("this is the name after clear: " + this.form.name.value)
-            this.form.declination.value = ""
-            this.form.declination.hasError = false
-
-            this.form.hours.value = ""
-            this.form.hours.hasError = false
-
-            this.form.minutes.value = ""
-            this.form.minutes.hasError = false
-
-            this.form.seconds.value = ""
-            this.form.seconds.hasError = false
-        },
+        updateBodies:function(){
+            console.log(this.bodies)
+            this.getCelestialBodies()
+        }
     },
     mounted: function(){
         this.getCelestialBodies();
